+++ includes/styles/a11y.php
@@ -5,7 +5,8 @@
  * @package    Additional Block Styles
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.4.0
  */
 
 namespace WebManDesign\ABS;
@@ -19,6 +20,8 @@ Register::add_style(
 		'label'  => _x( 'Accessibly hidden', 'Block style label.', 'additional-block-styles' ),
 		'blocks' => array(
 			'core/heading',
+			'core/site-title',
+			'core/site-tagline',
 		),
 	)
 );